backend/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
drf-yasg==1.20.0
elasticsearch==7.7.1
idna==2.9
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
kombu==4.6.9
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.17.5
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.0
scipy==1.4.1
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
